Moving all Inline types to it's own namespace
