+++ src/Telegram/Types/Inline/Query/Result/Gif.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace unreal4u\TelegramAPI\Telegram\Types\InlineQueryResult;
+namespace unreal4u\TelegramAPI\Telegram\Types\Inline\Query\Result;
 
-use unreal4u\TelegramAPI\Telegram\Types\InlineQueryResult;
+use unreal4u\TelegramAPI\Telegram\Types\Inline\Query\Result;
 
 /**
  * Represents a link to an animated GIF file. By default, this animated GIF file will be sent by the user with optional
@@ -14,7 +14,7 @@ use unreal4u\TelegramAPI\Telegram\Types\InlineQueryResult;
  *
  * @see https://core.telegram.org/bots/api#inlinequeryresultgif
  */
-class Gif extends InlineQueryResult
+class Gif extends Result
 {
     /**
      * Type of the result, must be gif
