+++ tests/Telegram/Methods/AnswerInlineQueryTest.php
@@ -4,7 +4,7 @@ namespace tests\Telegram\Methods;
 
 use tests\Mock\MockTgLog;
 use unreal4u\TelegramAPI\Telegram\Methods\AnswerInlineQuery;
-use unreal4u\TelegramAPI\Telegram\Types\InlineQueryResult\Article;
+use unreal4u\TelegramAPI\Telegram\Types\Inline\Query\Result\Article;
 
 class AnswerInlineQueryTest extends \PHPUnit_Framework_TestCase
 {
